third_party/WebKit/LayoutTests/http/tests/devtools/editor/text-editor-formatter.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../resources/editor-test.js"></script>
<script>
function codeSnippet() {
    return document.getElementById("codeSnippet").textContent;
}

function test() {
  var textEditor = SourcesTestRunner.createTestEditor();
  textEditor.setMimeType('text/javascript');
  textEditor.setReadOnly(false);
  textEditor.element.focus();

  function step2(result) {
    var codeLines = result.value;
    SourcesTestRunner.typeIn(textEditor, codeLines, step3);
  }

  function step3() {
    TestRunner.addResult('============ editor contents start ============');
    TestRunner.addResult(textEditor.text().replace(/ /g, '.'));
    TestRunner.addResult('============ editor contents end ============');
    TestRunner.completeTest();
  }

  TestRunner.evaluateInPage('codeSnippet();', step2);
}

</script>
</head>

<body onload="runTest();">
<p>
This test checks text editor javascript formatting.
</p>

<pre id="codeSnippet">/**
* Multi-line comment
*
*/
function foo(n) {
// one-line comment
function bar() {
return 42;D

var sum = 0;
for (var i = 0; i < n; ++i) {
for (var j = 0; j < n; ++j) {
sum += i + j;DD


if (sum > 1000) {
while (sum > 0) {
--sum;DDD
</pre>

</body>
</html>
