third_party/WebKit/LayoutTests/http/tests/devtools/editor/text-editor-formatter-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
This test checks text editor javascript formatting.

/**
* Multi-line comment
*
*/
function foo(n) {
// one-line comment
function bar() {
return 42;D

var sum = 0;
for (var i = 0; i < n; ++i) {
for (var j = 0; j < n; ++j) {
sum += i + j;DD


if (sum > 1000) {
while (sum > 0) {
--sum;DDD

============ editor contents start ============
/**
*.Multi-line.comment
*
*/
function.foo(n).{
....//.one-line.comment
....function.bar().{
........return.42;
....}

....var.sum.=.0;
....for.(var.i.=.0;.i.<.n;.++i).{
........for.(var.j.=.0;.j.<.n;.++j).{
............sum.+=.i.+.j;
........}
....}


....if.(sum.>.1000).{
........while.(sum.>.0).{
............--sum;
........}
....}
}

============ editor contents end ============
